goa: support unpacking zip files with bsdtar
Bsdtar, in contrast to unzip, allows to strip paths names (i.e.
`--strip-components=1`), which is needed to unpack software like
sqlite in a way that does not need adjusting of the src path in the
build step after updates.

If BSDTAR_OPT is set in the import file, unpack the zip file with bsdtar
with the options applied.

Fixes #125