fixup! recursively dereference symbolic links
Recursively dereference symbolic links by extending the path to resolve
symbolic links in the last element. The normalize trick is curtesy of:
https://wiki.tcl-lang.org/page/file+normalize

Prevents exporting api artifacts when the artifact links to a file in
the build directory which is itself a link to a file in the src
directory.

Fixes #130