set shared library linker flags automatically
Unlike Make, Meson does not offer a facility to override environment
variables in command line arguments and moreover, it does not support
referencing environment variables in `meson.build` files. This leaves us
in a position where presently Goa does not offer a built-in way to pass
in the correct linker flags for building shared libraries, since
arguments passed via `meson_args` will not expand referenced variables
in the spirit of `$LDLIBS_SHARED` nor can the flags generated by Goa be
referenced in a patched `meson.build` file.

Implement a heuristic that passes the correct linker flags to meson when all
build artifacts are shared libraries. This has the advantage that the
mechanism could easily be extended to obsolete manually overriding
`LDFLAGS` when building libraries with other build systems, at the expense
of only working for library-only archives.

Fixes #134